Split up remaining code in convert.go
- This leaves just the png -> txt conversion stuff in convert.go
- create metadata.go for the metadata files
- rename entries.go -> pokedex.go
- move other functions out into pokedex.go
