--- src/pokedex/entries.go
@@ -1,122 +1,0 @@
-package pokedex
-
-import (
-	"bufio"
-	"bytes"
-	"compress/gzip"
-	"encoding/gob"
-	"fmt"
-	"log"
-	"os"
-	"path"
-	"strings"
-)
-
-func check(e error) {
-	if e != nil {
-		log.Fatal(e)
-	}
-}
-
-type PokemonMetadata struct {
-	Categories       string
-	Name             string
-	JapaneseName     string
-	JapanesePhonetic string
-}
-
-func EntryFpath(subdir string, idx int) string {
-	return path.Join(subdir, fmt.Sprintf("%d.cow", idx))
-}
-
-func MetadataFpath(subdir string, idx int) string {
-	return path.Join(subdir, fmt.Sprintf("%d.metadata", idx))
-}
-func TokenizeName(name string) []string {
-	return strings.Split(name, "-")
-}
-
-func TokenizeCategories(categories string) []string {
-	return strings.Split(categories, "/")
-}
-
-type PokemonMatch struct {
-	Entry      *PokemonEntry
-	Categories []string
-}
-
-func Compress(data []byte) []byte {
-	var b bytes.Buffer
-	gz := gzip.NewWriter(&b)
-
-	_, err := gz.Write(data)
-	check(err)
-
-	err = gz.Close()
-	check(err)
-
-	return b.Bytes()
-}
-
-func Decompress(data []byte) []byte {
-	buf := bytes.NewBuffer(data)
-
-	reader, err := gzip.NewReader(buf)
-	check(err)
-
-	var resB bytes.Buffer
-
-	_, err = resB.ReadFrom(reader)
-	check(err)
-
-	return resB.Bytes()
-}
-
-func WriteStructToFile(data interface{}, fpath string) {
-	ostream, err := os.Create(fpath)
-	check(err)
-
-	writer := bufio.NewWriter(ostream)
-	enc := gob.NewEncoder(writer)
-	enc.Encode(data)
-	writer.Flush()
-	ostream.Close()
-}
-
-func ReadTrieFromBytes(data []byte) PokemonTrie {
-	buf := bytes.NewBuffer(data)
-	dec := gob.NewDecoder(buf)
-
-	d := &PokemonTrie{}
-
-	err := dec.Decode(&d)
-	check(err)
-
-	return *d
-}
-
-func ReadMetadataFromBytes(data []byte) PokemonMetadata {
-	buf := bytes.NewBuffer(data)
-	dec := gob.NewDecoder(buf)
-
-	d := &PokemonMetadata{}
-
-	err := dec.Decode(&d)
-	check(err)
-
-	return *d
-}
-
-func WriteBytesToFile(data []byte, fpath string, compress bool) {
-	ostream, err := os.Create(fpath)
-	check(err)
-
-	writer := bufio.NewWriter(ostream)
-	if compress {
-		writer.WriteString(string(Compress(data)))
-	} else {
-		writer.WriteString(string(data))
-	}
-	writer.Flush()
-	ostream.Close()
-}
